refactor: fetch 관련 코드와 쿼리 분리
